mgcp_network: use an uint16_t to store the port number
The comment in struct mgcp_rtp_end says that the RTCP port number has to
be stored in network byte order. Such a port number is 16 bit long, so
let's use an uint16_t instead of an int here.

Related: OS#6527

+++ include/osmocom/mgcp/mgcp_network.h
@@ -94,7 +94,7 @@ struct mgcp_rtp_end {
 	struct osmo_sockaddr addr;
 
 	/* in network byte order */
-	int rtcp_port;
+	uint16_t rtcp_port;
 
 	/* currently selected audio codec */
 	struct mgcp_rtp_codec *codec;
